refactor: 실패하는 테스트 구성
