+++ spring-mvc-2/initial/src/test/java/cholog/CRUDTest.java
@@ -11,7 +11,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class CRUDTest {
+class CRUDTest {
     @Test
     void create() {
         var response = RestAssured
